유저, 운동 기능 구현
- Spring Security 적용
- Auth 도메인 구현
- Exercise 도메인 구현
- Preference 도메인 Auth와 Exercise 연동

+++ src/main/java/com/healthmate/healthmate/domain/preference/controller/PreferenceController.java
@@ -39,8 +39,9 @@ public class PreferenceController {
 		return ResponseEntity.ok(preferenceService.getPreference(id));
 	}
 
+
 	@GetMapping
-	public ResponseEntity<List<PreferenceResponseDto>> getPreferencesByExercise(@RequestParam Integer exerciseId) {
+	public ResponseEntity<List<PreferenceResponseDto>> getPreferencesByExercise(@RequestParam Long exerciseId) {
 		return ResponseEntity.ok(preferenceService.getPreferencesByExercise(exerciseId));
 	}
