+++ src/main/java/com/healthmate/healthmate/domain/preference/dto/request/UpdatePreferenceRequestDto.java
@@ -3,7 +3,7 @@ package com.healthmate.healthmate.domain.preference.dto.request;
 import com.healthmate.healthmate.domain.preference.enums.PreferenceEnum;
 
 public record UpdatePreferenceRequestDto(
-	Integer exerciseId,
+	Long exerciseId,
 	PreferenceEnum preference
 ) {
 }
